# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumagram API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "api/"
        }
    ],
    "paths": {
...
        "/posts/{post}/comments": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Список комментариев",
                "operationId": "00a7052dfabc2d7ba47f1f5d5d62f176",
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Добавление комментария",
                "operationId": "98a3c321f84f38f3212cd81009382df5",
                "requestBody": {
                    "description": "Запрос",
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/users/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Профиль текущего пользователя",
                "operationId": "c26146dbc87681db03ebd0c6df994764",
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Редактирование профиля",
                "operationId": "f4a69737149a4e401e6d4b405bb88c2c",
                "requestBody": {
                    "description": "Запрос",
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}